feat(mapper): 更新FrameServiceRoleMapper和FrameServiceRoleServiceImpl中的服务ID类型为Long
- 将FrameServiceRoleMapper接口中的服务ID参数类型从Integer更改为Long，以支持更大的ID范围。
- 在FrameServiceRoleServiceImpl中相应地更新服务ID的处理逻辑，确保一致性。

这些改动提升了系统对服务ID的处理能力，确保了数据的一致性和准确性。

+++ datasophon-dao/src/main/java/com/datasophon/dao/mapper/FrameServiceRoleMapper.java
@@ -41,7 +41,7 @@ public interface FrameServiceRoleMapper extends BaseMapper<FrameServiceRoleEntit
          * @param serviceRoleType 服务角色类型，可为null
          * @return 服务角色列表
          */
-        default List<FrameServiceRoleEntity> selectByServiceIdsAndRoleType(@Param("serviceIds") List<Integer> serviceIds,
+        default List<FrameServiceRoleEntity> selectByServiceIdsAndRoleType(@Param("serviceIds") List<Long> serviceIds,
                         @Param("serviceRoleType") Integer serviceRoleType) {
                 QueryWrapper query = QueryWrapper.create()
                                 .where(FrameServiceRoleEntity::getServiceId).in(serviceIds);
@@ -91,7 +91,7 @@ public interface FrameServiceRoleMapper extends BaseMapper<FrameServiceRoleEntit
          * @param serviceIds 服务ID列表
          * @return 非MASTER角色列表
          */
-        default List<FrameServiceRoleEntity> selectNonMasterRoles(@Param("serviceIds") List<Integer> serviceIds) {
+        default List<FrameServiceRoleEntity> selectNonMasterRoles(@Param("serviceIds") List<Long> serviceIds) {
                 QueryWrapper query = QueryWrapper.create()
                                 .where(FrameServiceRoleEntity::getServiceRoleType).ne(1)
                                 .and(FrameServiceRoleEntity::getServiceId).in(serviceIds);
